src/genai_backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api==1.0.0
google-genai
python-dotenv
yt_dlp